morphism-openai reconcile #1: widen temperature f32 -> f64
Codec + openai.zig Options/Request temperature f32->f64 (Reader.readF64,
8-byte wire, codec.ts Writer.f64). Eliminates the precision divergence the
differential found (0.2 vs 0.20000000298). corpus full-case hex updated
(round-trip oracle validated); wasm rebuilt + re-embedded.

All 5 toOpenAI differential cases now pass as test() (flipped back from
test.failing). 151/151 core, 268/268 TS, tsc clean. fromOpenAI's 3 cases
remain test.failing pending reconcile #2 (Zig Response envelope leniency).

+++ core/src/morphisms/openai/openai.zig
@@ -26,7 +26,7 @@ const luv = @import("../luv/luv.zig");
 pub const Options = struct {
     model: []const u8,
     max_tokens: ?u32 = null,
-    temperature: ?f32 = null,
+    temperature: ?f64 = null,
     stream: bool = false,
     tools: []const luv.Tool = &.{},
 };
@@ -65,7 +65,7 @@ pub const Request = struct {
     model: []const u8,
     messages: []const RequestMessage,
     max_tokens: ?u32 = null,
-    temperature: ?f32 = null,
+    temperature: ?f64 = null,
     stream: ?bool = null,
     tools: ?[]const ToolDef = null,
 };
